Fixes suggested by Trevor.
