+++ gateway-js/src/__tests__/integration/networkRequests.test.ts
@@ -11,7 +11,7 @@ import {
   mockSupergraphSdlRequestSuccess,
   mockSupergraphSdlRequest,
   mockApolloConfig,
-  mockCloudConfigUrl,
+  mockCloudConfigUrl1,
   mockSupergraphSdlRequestIfAfter,
   mockSupergraphSdlRequestSuccessIfAfter,
 } from './nockMocks';
@@ -105,7 +105,7 @@ it('Fetches Supergraph SDL from remote storage', async () => {
 
   gateway = new ApolloGateway({
     logger,
-    schemaConfigDeliveryEndpoints: [mockCloudConfigUrl],
+    uplinkEndpoints: [mockCloudConfigUrl1],
   });
 
   await gateway.load(mockApolloConfig);
@@ -115,7 +115,7 @@ it('Fetches Supergraph SDL from remote storage', async () => {
 
 it('Fetches Supergraph SDL from remote storage using a configured env variable', async () => {
   cleanUp = mockedEnv({
-    APOLLO_SCHEMA_CONFIG_DELIVERY_ENDPOINT: mockCloudConfigUrl,
+    APOLLO_SCHEMA_CONFIG_DELIVERY_ENDPOINT: mockCloudConfigUrl1,
   });
   mockSupergraphSdlRequestSuccess();
 
@@ -149,7 +149,7 @@ it('Updates Supergraph SDL from remote storage', async () => {
 
   gateway = new ApolloGateway({
     logger,
-    schemaConfigDeliveryEndpoints: [mockCloudConfigUrl],
+    uplinkEndpoints: [mockCloudConfigUrl1],
   });
   // eslint-disable-next-line
   // @ts-ignore for testing purposes, a short pollInterval is ideal so we'll override here
@@ -169,8 +169,8 @@ describe('Supergraph SDL update failures', () => {
 
     gateway = new ApolloGateway({
       logger,
-      schemaConfigDeliveryEndpoints: [mockCloudConfigUrl],
-      schemaDeliveryMaxRetries: 0
+      uplinkEndpoints: [mockCloudConfigUrl1],
+      uplinkMaxRetries: 0
     });
 
     await expect(
@@ -198,8 +198,8 @@ describe('Supergraph SDL update failures', () => {
 
     gateway = new ApolloGateway({
       logger,
-      schemaConfigDeliveryEndpoints: [mockCloudConfigUrl],
-      schemaDeliveryMaxRetries: 0
+      uplinkEndpoints: [mockCloudConfigUrl1],
+      uplinkMaxRetries: 0
     });
 
     // eslint-disable-next-line
@@ -233,8 +233,8 @@ describe('Supergraph SDL update failures', () => {
 
     gateway = new ApolloGateway({
       logger,
-      schemaConfigDeliveryEndpoints: [mockCloudConfigUrl],
-      schemaDeliveryMaxRetries: 0
+      uplinkEndpoints: [mockCloudConfigUrl1],
+      uplinkMaxRetries: 0
     });
     // eslint-disable-next-line
     // @ts-ignore for testing purposes, a short pollInterval is ideal so we'll override here
@@ -272,7 +272,7 @@ describe('Supergraph SDL update failures', () => {
 
     gateway = new ApolloGateway({
       logger,
-      schemaConfigDeliveryEndpoints: [mockCloudConfigUrl],
+      uplinkEndpoints: [mockCloudConfigUrl1],
     });
     // eslint-disable-next-line
     // @ts-ignore for testing purposes, a short pollInterval is ideal so we'll override here
@@ -303,7 +303,7 @@ describe('Supergraph SDL update failures', () => {
 
     gateway = new ApolloGateway({
       logger,
-      schemaConfigDeliveryEndpoints: [mockCloudConfigUrl],
+      uplinkEndpoints: [mockCloudConfigUrl1],
     });
     // eslint-disable-next-line
     // @ts-ignore for testing purposes, a short pollInterval is ideal so we'll override here
@@ -348,7 +348,7 @@ it('Rollsback to a previous schema when triggered', async () => {
 
   gateway = new ApolloGateway({
     logger,
-    schemaConfigDeliveryEndpoints: [mockCloudConfigUrl],
+    uplinkEndpoints: [mockCloudConfigUrl1],
   });
   // eslint-disable-next-line
   // @ts-ignore for testing purposes, a short pollInterval is ideal so we'll override here
@@ -434,7 +434,7 @@ describe('Downstream service health checks', () => {
       gateway = new ApolloGateway({
         serviceHealthCheck: true,
         logger,
-        schemaConfigDeliveryEndpoints: [mockCloudConfigUrl],
+        uplinkEndpoints: [mockCloudConfigUrl1],
       });
       // eslint-disable-next-line
       // @ts-ignore for testing purposes, a short pollInterval is ideal so we'll override here
@@ -458,7 +458,7 @@ describe('Downstream service health checks', () => {
       gateway = new ApolloGateway({
         serviceHealthCheck: true,
         logger,
-        schemaConfigDeliveryEndpoints: [mockCloudConfigUrl],
+        uplinkEndpoints: [mockCloudConfigUrl1],
       });
 
       // This is the ideal, but our version of Jest has a bug with printing error snapshots.
@@ -519,7 +519,7 @@ describe('Downstream service health checks', () => {
       gateway = new ApolloGateway({
         serviceHealthCheck: true,
         logger,
-        schemaConfigDeliveryEndpoints: [mockCloudConfigUrl],
+        uplinkEndpoints: [mockCloudConfigUrl1],
       });
       // eslint-disable-next-line @typescript-eslint/ban-ts-comment
       // @ts-ignore for testing purposes, a short pollInterval is ideal so we'll override here
@@ -563,7 +563,7 @@ describe('Downstream service health checks', () => {
       gateway = new ApolloGateway({
         serviceHealthCheck: true,
         logger,
-        schemaConfigDeliveryEndpoints: [mockCloudConfigUrl],
+        uplinkEndpoints: [mockCloudConfigUrl1],
       });
       // eslint-disable-next-line
       // @ts-ignore for testing purposes, a short pollInterval is ideal so we'll override here
